Fix compiler warning [-Wpointer-to-int-cast].
* DocFormats/core/src/lib/DFAllocator.c:
  (DFAllocatorAlloc):
     Change cast of (unsigned long long)mem to (unsigned long) to
     remove compiler warning.

+++ DocFormats/core/src/lib/DFAllocator.c
@@ -77,6 +77,6 @@ void *DFAllocatorAlloc(DFAllocator *alc, size_t size)
     char *mem = block->mem + block->used;
     block->used += size;
     assert(block->used <= block->size);
-    assert((((unsigned long long)mem) % 8) == 0);
+    assert((((unsigned long)mem) % 8) == 0);
     return mem;
 }
